Infer epimutation clocks in number of cell divisions and fitness of + cells with respect to - cells.

A PEPI fit with fitness inference is returned.

fit_s(
  x,
  path_to_model,
  cmdstan_path,
  threshold = 0.1,
  ndraws = 1000,
  init = NULL,
  seed = 45,
  mu = 1e-07,
  l = 2.7 * 10^9,
  ms = -0.5,
  sigma = 0.5,
  k = 100
)

Arguments

x

PEPI object containing fitness and epimutation clocks inference

path_to_model

String specifying the path where we want to save the stan model for a given depth

cmdstan_path

String specifying the path to cmdstan folder

threshold

Threshold for tree pruning

ndraws

Number of draws from the posterior

init

List of initialization parameters

seed

Seed of the computation

mu

Mutation rate per division per bp per allele

l

length of the genome

ms

Mean of lognormal prior for s @param sigma Sigma parameter of lognormal prior for s

k

Number of trials for the beta prior on cluster centroids

Value

PEPI object

Examples

fit_s(x,path_to_model = "models",cmdstan_path = "my_cmdstan/",threshold = 0.1,
ndraws = 1000,init = NULL,seed = 45,
mu = 1e-7,l = 2.7*10^9,ms = -0.5,sigma = 0.5,k = 100)
#> Warning: 'models' already exists
#> Warning: Path not set. Can't find directory: my_cmdstan/
#> Error in fit_s(x, path_to_model = "models", cmdstan_path = "my_cmdstan/",     threshold = 0.1, ndraws = 1000, init = NULL, seed = 45, mu = 1e-07,     l = 2.7 * 10^9, ms = -0.5, sigma = 0.5, k = 100): object 'x' not found