Fit a multivariate vaf spectrum with epigenetic tree model.

A PEPI fit with tree inference is returned.

fit_tree(
  x,
  path_to_model,
  cmdstan_path,
  max_depth = 2,
  ndraws = 1000,
  init = NULL,
  seed = 15,
  mu = 1e-07,
  l = 2.7 * 10^9,
  rho_n = 1,
  rho_p = 1,
  nu_t = 0.1,
  qt = 10000,
  rate_n = 0.001,
  qn = 10000,
  rate_p = 0.001,
  qp = 10000,
  k = 10000,
  gamma = 150
)

Arguments

x

PEPI object containing VAF multivariate spectrum

path_to_model

String specifying the path where we wannt to save the stan model for a given depth

cmdstan_path

String specifying the path to cmdstan folder

max_depth

Maximum number of levels

ndraws

Number of draws from the posterior

init

List of initialization parameters

seed

Seed of the computation

mu

Mutation rate per division per bp per allele

l

length of the genome

rho_n

Purity of - sample @param rho_p Purity of + sample

nu_t

Mean of the beta prior on fraction of truncal mutations

qt

Number of trials of the beta prior on fraction of truncal mutations

rate_n

Mean of the beta prior on the epimutation rate from - to +

qn

Number of trials for the beta prior on the epimutation rate from - to +

rate_p

Mean of the beta prior on the epimutation rate from + to -

qp

Number of trials for the beta prior on the epimutation rate from - to +

k

Number of trials for the beta prior on cluster centroids

gamma

Concentration of a Dirichlet distribution to split mutations at any node

Value

PEPI object

Examples

fit_tree(x,path_to_model = "models",cmdstan_path = "my_cmdstan/",
max_depth = 2,ndraws = 1000,init = NULL,seed = 15,
mu = 1e-7,l = 2.7*10^9,rho_n = 1,rho_p = 1,nu_t = 0.1,
qt = 1e4,rate_n = 1e-3,qn = 1e4,rate_p = 1e-3,
qp = 1e4,k = 1e4,gamma = 150)
#> Warning: 'models' already exists
#> Warning: Path not set. Can't find directory: my_cmdstan/
#> Error in fit_tree(x, path_to_model = "models", cmdstan_path = "my_cmdstan/",     max_depth = 2, ndraws = 1000, init = NULL, seed = 15, mu = 1e-07,     l = 2.7 * 10^9, rho_n = 1, rho_p = 1, nu_t = 0.1, qt = 10000,     rate_n = 0.001, qn = 10000, rate_p = 0.001, qp = 10000, k = 10000,     gamma = 150): object 'x' not found