generative models Bla bla

Generate a sample tree with input epimutation rates, fraction of truncal mutations and fitness advantage between + and -.

A tree with number of mutations per nodes is generated.

generate_nodes(
  M,
  gamma,
  at,
  bt,
  s,
  rate_minus,
  rate_plus,
  mu,
  len,
  max_depth = 3
)

Arguments

M

total number of mutations

gamma

parameter of a beta distribution controlling how we split mutations at any node

at

number of successes of a beta distribution controlling amount of truncal mutations

bt

number of fails of a beta distribution controlling amount of truncal mutations

s

Fitness advantage of + cells with respect to - cells

rate_minus

Average number of epimutation from - to + per mutations

rate_plus

Average number of epimutation from + to - per mutations

mu

Mutation rate per division per bp per allele

len

length of the genome

max_depth

maximal level of the tree

Value

sample tree

Examples

generate_nodes(6000,100,10,80,0.2,1e-3,1e-3,1e-7,2.7*10^9,3)
#> Error in data.frame(node = "-", level = 0, delta_m = M, phi = 1, p_switch = 1,     nu = rbeta(1, at, bt), rate = rate_minus, s = 0) %>% mutate(m = nu *     M): could not find function "%>%"